refactor(macros): simplify internal subsystem layout
