+++ statum-macros/src/validators.rs
@@ -1,60 +1,33 @@
-use quote::{ToTokens, format_ident, quote};
-use std::collections::HashMap;
-use syn::{Generics, Ident, ItemImpl, Path, Type, parse_macro_input};
+//! `#[validators]` subsystem: resolve target machines, validate signatures, and emit rebuild helpers.
 
+use quote::{ToTokens, format_ident, quote};
+use syn::{ItemImpl, Path, parse_macro_input};
+
+use crate::contracts::ValidatorContract;
 use crate::diagnostics::{DiagnosticMessage, compile_error_at};
-use crate::VariantInfo;
 use crate::machine::{
-    builder_generics, extra_generics, extra_type_arguments_tokens, generic_argument_tokens,
-    machine_type_with_state,
+    builder_generics, extra_generics, generic_argument_tokens, machine_type_with_state,
 };
 
+mod contract;
 mod emission;
+mod plan;
 mod resolution;
 mod signatures;
 mod type_equivalence;
 
-use emission::{
-    BatchBuilderContext, ValidatorCheckContext, batch_builder_implementation,
-    generate_validator_check, generate_validator_report_check, inject_machine_fields,
+use contract::{
+    CollectValidatorContext, IntoMachineBuilderContext, build_validator_contract,
+    machine_scoped_item_path, qualify_machine_field_types,
 };
+use emission::{
+    BatchBuilderContext, batch_builder_implementation, inject_machine_fields,
+};
+use plan::collect_validator_checks;
 use resolution::{
     resolve_machine_metadata, resolve_state_enum_info, resolve_validator_machine_attr,
     validate_validator_coverage,
 };
-use signatures::{
-    ValidatorDiagnosticContext, validate_validator_return_type, validate_validator_signature,
-    validator_state_name_from_ident,
-};
-
-struct VariantSpec {
-    variant_name: String,
-    has_state_data: bool,
-    expected_ok_type: Type,
-}
-
-struct CollectValidatorContext<'a> {
-    machine_path: &'a Path,
-    machine_module_path: &'a Path,
-    machine_generics: &'a Generics,
-    field_names: &'a [Ident],
-    persisted_type_display: &'a str,
-    machine_name: &'a str,
-    state_enum_name: &'a str,
-}
-
-struct IntoMachineBuilderContext<'a> {
-    builder_ident: &'a Ident,
-    struct_ident: &'a Type,
-    machine_generics: &'a Generics,
-    machine_state_ty: &'a proc_macro2::TokenStream,
-    field_names: &'a [Ident],
-    field_types: &'a [Type],
-    validator_checks: &'a [proc_macro2::TokenStream],
-    validator_report_checks: &'a [proc_macro2::TokenStream],
-    async_token: &'a proc_macro2::TokenStream,
-    machine_vis: &'a syn::Visibility,
-}
 
 pub fn parse_validators(
     attr: proc_macro::TokenStream,
@@ -68,11 +41,6 @@ pub fn parse_validators(
         Ok(attr) => attr,
         Err(err) => return err.into(),
     };
-    let machine_ident = machine_attr.machine_ident.clone();
-    let machine_name = machine_attr.machine_name.clone();
-    let machine_attr_display = machine_attr.attr_display.clone();
-    let machine_module_path =
-        machine_support_module_path(&machine_attr.machine_path, &machine_name);
 
     let machine_metadata = match resolve_machine_metadata(module_path, &machine_attr) {
         Ok(metadata) => metadata,
@@ -101,41 +69,45 @@ pub fn parse_validators(
         Err(err) => return err.into(),
     };
 
+    let contract = build_validator_contract(
+        &machine_attr,
+        machine_metadata.clone(),
+        parsed_machine,
+        &parsed_fields,
+        state_enum_info,
+        &persisted_type_display,
+    );
+    let ValidatorContract {
+        resolved_machine,
+        state_enum,
+        persisted_type_display,
+        machine_attr_display,
+    } = contract;
+
     let validator_coverage = match validate_validator_coverage(
         &item_impl,
-        &state_enum_info,
+        &state_enum.enum_info,
         &persisted_type_display,
         &machine_attr_display,
-        &machine_name,
+        &resolved_machine.machine_name,
     ) {
         Ok(()) => quote! {},
         Err(err) => return err.into(),
     };
 
-    let field_names = parsed_fields
-        .iter()
-        .map(|(ident, _)| ident.clone())
-        .collect::<Vec<_>>();
-    let field_types = parsed_fields
-        .iter()
-        .map(|(_, ty)| ty.clone())
-        .collect::<Vec<_>>();
-    let machine_extra_ty_args = extra_type_arguments_tokens(&parsed_machine.generics);
-    let machine_state_ty = quote! { #machine_module_path::SomeState #machine_extra_ty_args };
-
     let collect_context = CollectValidatorContext {
-        machine_path: &machine_attr.machine_path,
-        machine_module_path: &machine_module_path,
-        machine_generics: &parsed_machine.generics,
-        field_names: &field_names,
+        machine_path: &resolved_machine.machine_path,
+        machine_module_path: &resolved_machine.machine_module_path,
+        machine_generics: resolved_machine.machine_generics(),
+        field_names: &resolved_machine.field_names,
         persisted_type_display: &persisted_type_display,
-        machine_name: &machine_name,
-        state_enum_name: &state_enum_info.name,
+        machine_name: &resolved_machine.machine_name,
+        state_enum_name: &resolved_machine.state_enum_name,
     };
 
     let (validator_checks, validator_report_checks, has_async) = match collect_validator_checks(
         &item_impl,
-        &state_enum_info.variants,
+        &state_enum.variants,
         &collect_context,
     ) {
         Ok(result) => result,
@@ -143,13 +115,13 @@ pub fn parse_validators(
     };
 
     if item_impl.items.is_empty() {
-        let expected_methods = state_enum_info
+        let expected_methods = state_enum
             .variants
             .iter()
             .map(|variant| format!("is_{}", crate::to_snake_case(&variant.name)))
             .collect::<Vec<_>>()
             .join(", ");
-        let state_enum_name = state_enum_info.name.clone();
+        let state_enum_name = state_enum.name.clone();
         let message = DiagnosticMessage::new(format!(
             "`#[validators({machine_attr_display})]` on `impl {persisted_type_display}` must define at least one validator method."
         ))
@@ -160,7 +132,7 @@ pub fn parse_validators(
         return compile_error_at(proc_macro2::Span::call_site(), &message).into();
     }
 
-    let machine_vis = parsed_machine.vis.clone();
+    let machine_vis = resolved_machine.parsed_machine.vis.clone();
 
     let async_token = if has_async {
         quote! { async }
@@ -169,38 +141,40 @@ pub fn parse_validators(
     };
 
     let batch_builder_impl = batch_builder_implementation(BatchBuilderContext {
-        machine_ident: &machine_ident,
-        machine_module_path: &machine_module_path,
-        machine_generics: &parsed_machine.generics,
+        machine_ident: &resolved_machine.machine_ident,
+        machine_module_path: &resolved_machine.machine_module_path,
+        machine_generics: resolved_machine.machine_generics(),
         struct_ident,
-        machine_state_ty: &machine_state_ty,
-        field_names: &field_names,
-        field_types: &field_types,
+        machine_state_ty: &resolved_machine.machine_state_ty,
+        field_names: &resolved_machine.field_names,
+        field_types: &resolved_machine.field_types,
         async_token: async_token.clone(),
         machine_vis: machine_vis.clone(),
     });
 
-    let into_machine_builder_ident = format_ident!("__Statum{}IntoMachine", machine_ident);
-    let into_machines_builder_ident = format_ident!("__Statum{}IntoMachines", machine_ident);
+    let into_machine_builder_ident =
+        format_ident!("__Statum{}IntoMachine", resolved_machine.machine_ident);
+    let into_machines_builder_ident =
+        format_ident!("__Statum{}IntoMachines", resolved_machine.machine_ident);
     let into_machine_builder_impl = generate_into_machine_builder(IntoMachineBuilderContext {
         builder_ident: &into_machine_builder_ident,
         struct_ident,
-        machine_generics: &parsed_machine.generics,
-        machine_state_ty: &machine_state_ty,
-        field_names: &field_names,
-        field_types: &field_types,
+        machine_generics: resolved_machine.machine_generics(),
+        machine_state_ty: &resolved_machine.machine_state_ty,
+        field_names: &resolved_machine.field_names,
+        field_types: &resolved_machine.field_types,
         validator_checks: &validator_checks,
         validator_report_checks: &validator_report_checks,
         async_token: &async_token,
         machine_vis: &machine_vis,
     });
-    let into_machine_extra_generics = extra_generics(&parsed_machine.generics);
-    let slot_storage_idents = (0..field_names.len())
+    let into_machine_extra_generics = extra_generics(resolved_machine.machine_generics());
+    let slot_storage_idents = (0..resolved_machine.field_names.len())
         .map(|idx| format_ident!("__statum_slot_{}", idx))
         .collect::<Vec<_>>();
     let (into_machine_method_generics, _, into_machine_method_where_clause) =
         into_machine_extra_generics.split_for_impl();
-    let into_machine_slot_defaults = (0..field_names.len())
+    let into_machine_slot_defaults = (0..resolved_machine.field_names.len())
         .map(|_| quote! { false })
         .collect::<Vec<_>>();
     let into_machines_builder_ty_generics = generic_argument_tokens(
@@ -218,15 +192,15 @@ pub fn parse_validators(
         Some(quote! { '__statum_row }),
         &into_machine_slot_defaults,
     );
-    let uninitialized_state_ident =
-        format_ident!("Uninitialized{}", state_enum_info.name);
+    let uninitialized_state_ident = format_ident!("Uninitialized{}", state_enum.name);
     let uninitialized_state_path =
         machine_scoped_item_path(&machine_attr.machine_path, &uninitialized_state_ident);
     let uninitialized_machine_ty = machine_type_with_state(
         quote! { #machine_path },
-        &parsed_machine.generics,
+        resolved_machine.machine_generics(),
         quote! { #uninitialized_state_path },
     );
+    let machine_module_path = &resolved_machine.machine_module_path;
 
     let machine_builder_impl = quote! {
         #[allow(unused_imports)]
@@ -277,150 +251,6 @@ pub fn parse_validators(
     };
 
     expanded.into()
-}
-
-fn collect_validator_checks(
-    item_impl: &ItemImpl,
-    variants: &[VariantInfo],
-    context: &CollectValidatorContext<'_>,
-) -> Result<
-    (
-        Vec<proc_macro2::TokenStream>,
-        Vec<proc_macro2::TokenStream>,
-        bool,
-    ),
-    proc_macro2::TokenStream,
-> {
-    let mut checks = Vec::new();
-    let mut report_checks = Vec::new();
-    let mut has_async = false;
-    let receiver = quote! { __statum_persisted };
-    let (variant_specs, variant_by_name) = build_variant_lookup(variants)?;
-    let emission_context = ValidatorCheckContext {
-        machine_path: context.machine_path,
-        machine_module_path: context.machine_module_path,
-        machine_generics: context.machine_generics,
-        field_names: context.field_names,
-        receiver: &receiver,
-    };
-
-    for item in &item_impl.items {
-        let syn::ImplItem::Fn(func) = item else {
-            continue;
-        };
-
-        let Some(state_name) = validator_state_name_from_ident(&func.sig.ident) else {
-            continue;
-        };
-        let Some(spec_idx) = variant_by_name.get(&state_name) else {
-            continue;
-        };
-        let spec = &variant_specs[*spec_idx];
-        let diagnostic_context = ValidatorDiagnosticContext {
-            persisted_type_display: context.persisted_type_display,
-            machine_name: context.machine_name,
-            state_enum_name: context.state_enum_name,
-            variant_name: &spec.variant_name,
-            machine_fields: context.field_names,
-            expected_ok_type: &spec.expected_ok_type,
-        };
-        validate_validator_signature(func, &diagnostic_context)?;
-        let return_kind =
-            validate_validator_return_type(func, &spec.expected_ok_type, &diagnostic_context)?;
-
-        if func.sig.asyncness.is_some() {
-            has_async = true;
-        }
-        checks.push(generate_validator_check(
-            &emission_context,
-            &spec.variant_name,
-            spec.has_state_data,
-            func.sig.asyncness.is_some(),
-        ));
-        report_checks.push(generate_validator_report_check(
-            &emission_context,
-            &spec.variant_name,
-            spec.has_state_data,
-            return_kind,
-            func.sig.asyncness.is_some(),
-        ));
-    }
-
-    Ok((checks, report_checks, has_async))
-}
-
-fn machine_support_module_path(machine_path: &Path, machine_name: &str) -> Path {
-    let mut support_path = machine_path.clone();
-    if let Some(last_segment) = support_path.segments.last_mut() {
-        last_segment.ident = format_ident!("{}", crate::to_snake_case(machine_name));
-    }
-    support_path
-}
-
-fn machine_scoped_item_path(machine_path: &Path, item_ident: &Ident) -> Path {
-    let mut scoped_path = machine_path.clone();
-    if let Some(last_segment) = scoped_path.segments.last_mut() {
-        last_segment.ident = item_ident.clone();
-    }
-    scoped_path
-}
-
-fn qualify_machine_field_types(
-    parsed_fields: &[(Ident, Type)],
-    machine_path: &Path,
-) -> Vec<(Ident, Type)> {
-    parsed_fields
-        .iter()
-        .map(|(ident, field_ty)| {
-            (
-                ident.clone(),
-                qualify_machine_scoped_type(field_ty, machine_path),
-            )
-        })
-        .collect()
-}
-
-fn qualify_machine_scoped_type(field_ty: &Type, machine_path: &Path) -> Type {
-    let Type::Path(type_path) = field_ty else {
-        return field_ty.clone();
-    };
-    if type_path.qself.is_some()
-        || type_path.path.leading_colon.is_some()
-        || type_path.path.segments.len() != 1
-    {
-        return field_ty.clone();
-    }
-
-    let Some(segment) = type_path.path.segments.last() else {
-        return field_ty.clone();
-    };
-    let mut qualified = machine_scoped_item_path(machine_path, &segment.ident);
-    if let Some(last_segment) = qualified.segments.last_mut() {
-        last_segment.arguments = segment.arguments.clone();
-    }
-
-    syn::parse_quote!(#qualified)
-}
-
-fn build_variant_lookup(
-    variants: &[VariantInfo],
-) -> Result<(Vec<VariantSpec>, HashMap<String, usize>), proc_macro2::TokenStream> {
-    let mut specs = Vec::with_capacity(variants.len());
-    let mut variant_by_name = HashMap::with_capacity(variants.len() * 2);
-
-    for variant in variants {
-        let state_data_type = variant.parse_data_type()?;
-        specs.push(VariantSpec {
-            variant_name: variant.name.clone(),
-            has_state_data: state_data_type.is_some(),
-            expected_ok_type: state_data_type.unwrap_or_else(|| syn::parse_quote!(())),
-        });
-        let idx = specs.len() - 1;
-        variant_by_name.insert(variant.name.clone(), idx);
-        variant_by_name.insert(crate::to_snake_case(&variant.name), idx);
-    }
-
-    Ok((specs, variant_by_name))
 }
 
 fn generate_into_machine_builder(context: IntoMachineBuilderContext<'_>) -> proc_macro2::TokenStream {
